+++ tests/api/test_page.py
@@ -235,3 +235,27 @@ def test_bulk_publish(
         page.refresh_from_db()
     assert content['data']['pageBulkPublish']['count'] == len(page_list)
     assert all(page.is_published for page in page_list)
+
+
+def test_bulk_unpublish(
+        staff_api_client, page_list,
+        permission_manage_pages):
+    assert all(page.is_published for page in page_list)
+
+    query = """
+        mutation unPublishManyPages($ids: [ID]!) {
+            pageBulkUnpublish(ids: $ids) {
+                count
+            }
+        }
+    """
+    variables = {'ids': [
+        graphene.Node.to_global_id('Page', page.id)
+        for page in page_list]}
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_pages])
+    content = get_graphql_content(response)
+    for page in page_list:
+        page.refresh_from_db()
+    assert content['data']['pageBulkUnpublish']['count'] == len(page_list)
+    assert not any(page.is_published for page in page_list)